+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+- Shift+Return hotkey to explicitly disable minimizing when making a selection
+
 ## [0.24.1] - 2019-09-08
 ### Changed
 - Tab hotkeys changed from `Alt+<number>` to `Ctrl+<number>` on non-Linux for consistency with other applications